fixed local storage saving repeats

+++ assets/js/DevNana PERS FOLDER/films.js
@@ -37,13 +37,13 @@ function search(event) {
         resultButton.attr("data-id", link[i].imdbid);
         resultButton.addClass("summary");
         resultButton.css({
-          'width': 182 + 'px',
-          'height': 268 + 'px',
-          'background-size': 'cover',
-          'background-position': 'center center',
-          'padding': '10px',
-          'margin': '10px',
-          'border-radius': '10px',
+          width: 182 + "px",
+          height: 268 + "px",
+          "background-size": "cover",
+          "background-position": "center center",
+          padding: "10px",
+          margin: "10px",
+          "border-radius": "10px",
         });
 
         // if image url doesn't work loads a replacement image
@@ -53,8 +53,6 @@ function search(event) {
             event.target.backgroundImage = `url("https://t4.ftcdn.net/jpg/02/97/01/65/360_F_297016511_NWrJG1s3mpyjqD3hwdKidfYsvhEnrPm4.jpg")`;
             event.onerror = null;
           });
-          
-          
         });
 
         $("#columnDisplay").append(resultButton);
@@ -104,13 +102,13 @@ function movieOption(event) {
         resultButton.attr("data-id", randomLink[i].imdbid);
         resultButton.addClass("summary");
         resultButton.css({
-          'width': 182 + 'px',
-          'height': 268 + 'px',
-          'background-size': 'cover',
-          'background-position': 'center center',
-          'padding': '10px',
-          'margin': '10px',
-          'border-radius': '10px',
+          width: 182 + "px",
+          height: 268 + "px",
+          "background-size": "cover",
+          "background-position": "center center",
+          padding: "10px",
+          margin: "10px",
+          "border-radius": "10px",
         });
 
         // if image url doesn't work loads a replacement image
@@ -120,7 +118,6 @@ function movieOption(event) {
             event.target.backgroundImage = `url("https://t4.ftcdn.net/jpg/02/97/01/65/360_F_297016511_NWrJG1s3mpyjqD3hwdKidfYsvhEnrPm4.jpg")`;
             event.onerror = null;
           });
-          
         });
 
         $("#columnDisplay").append(resultButton);
@@ -157,17 +154,29 @@ function createButton(value) {
   const existingSearchData = localStorage.getItem("searches");
   const searches = existingSearchData ? JSON.parse(existingSearchData) : []; // checking to see if there is anything in local storage if so converts it into a javascript object if not places an empty array
 
-  if (searches.length >= 5) {
-    // checking the number of propertys in local storage if equal or greater than five
-    searches.shift(); // if greater than or equal to removes first property
+  let alreadyExists = false;
+
+  for (let i = 0; i < searches.length; i++) {
+    if (searches[i].text === buttonData.text) {
+      alreadyExists = true;
+      break;
+    }
+  } // checking to see if the button already exists
+
+  if (!alreadyExists) {
+    //makes only happens if the button doesn't already exist
+    if (searches.length >= 5) {
+      // checking the number of propertys in local storage if equal or greater than five
+      searches.shift(); // if greater than or equal to removes first property
+    }
+
+    searches.push(buttonData); // pushes the value entered to the object
+
+    const searchesJson = JSON.stringify(searches); // stringifys the key and value in the object
+    localStorage.setItem("searches", searchesJson); // saves to local storage
+
+    renderButtons(); // Call renderButtons after creating the new button
   }
-
-  searches.push(buttonData); // pushes the value entered to the object
-
-  const searchesJson = JSON.stringify(searches); // stringifys the key and value in the object
-  localStorage.setItem("searches", searchesJson); // saves to local storage
-
-  renderButtons(); // Call renderButtons after creating the new button
 }
 
 function renderButtons() {
@@ -208,13 +217,13 @@ function renderButtons() {
             resultButton.attr("data-id", link[i].imdbid);
             resultButton.addClass("summary");
             resultButton.css({
-              'width': 182 + 'px',
-              'height': 268 + 'px',
-              'background-size': 'cover',
-              'background-position': 'center center',
-              'padding': '10px',
-              'margin': '10px',
-              'border-radius': '10px',
+              width: 182 + "px",
+              height: 268 + "px",
+              "background-size": "cover",
+              "background-position": "center center",
+              padding: "10px",
+              margin: "10px",
+              "border-radius": "10px",
             });
 
             // if image url doesn't work loads a replacement image
@@ -224,7 +233,6 @@ function renderButtons() {
                 event.target.backgroundImage = `url("https://t4.ftcdn.net/jpg/02/97/01/65/360_F_297016511_NWrJG1s3mpyjqD3hwdKidfYsvhEnrPm4.jpg")`;
                 event.onerror = null;
               });
-              
             });
 
             $("#columnDisplay").append(resultButton);
